os/binfmt: Fix memfault during Common binary Update test
During update test, we try to unload and reload all binaries. In case
of app binary, we disbale the MPU regions based on the MPU regs stored
in the tcb of apps. But, in case of Common binary, the MPU regions are
not stored anywhere. As a result, the MPU for common binary is not
disabled. So, at the time of reloading common binary, there is a memfault
while trying to copy code from binary to the read-only area of the code
section.

Issue is fixed by storing the common binary MPU registers in the binary
data structure and disabling the regions at the time of unload module.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>