kernel/binary_manager: Update bootparam when the binary not written in bootparam is loaded
Binary manager basically loads binaries based on boot parameters.
But if the binary written in bootparam is invalid, it should check and load the binary in another partition.
Currently it doesn't work as above.

This commit fixes to check another partition in case of invalid binary in a partition based on bootparam.