src/armv7-m,armv8-m: Modify to get the correct assert pc
During the faults and assert failure in app space, g_assert_location is set
to appropriate value to indicate the location of the crash and is used during
up_assert. In case of app SW assert, the crash location is obtained from
R14 of the interrupt context and in case of HW fault in app, it is obtained
from R15 of the interrupt context saved on the stack. In the case of kernel
asserts, we get the assert pc from the return address (current LR value) in
up_assert.

Signed-off-by: Abhishek Akkabathula <a.akkabathul@samsung.com>