net/lwip: fix socket descriptor is not shared to task group
netconn was refered by pid not task group. So if a task A which created
socket descriptor was terminated then the other task which was created
by task A uses socket with accept then it generates crash because
netconn contains the pid of task which create socket. So it's safe to
contains task_group_s instead of pid to netconn.