os/board/rtl8721csm: Fix uart data loss while flash write
This PR contains modifications on uart api and flash api.

For the uart api update:
tx_level is added to ensure 16 bytes of data is transmitted for each uart tx interrupt.
In this way, the total number of interrupts can be reduced as well as the time used for the transmission.

For the flash api update:
FLASH_Write_Lock is applied only before and after the flash write functions to reduce the overall period of FLASH_Write_Lock.