fixup! net/lwip: support window scale option
CONFIG_NET_WND_SCALE is a bool type so that it can't have a range syntax.
This causes menuconfig warnings as below:

net/lwip/configs/tcp/Kconfig:23:warning: range is only allowed for int or hex symbols
net/lwip/configs/tcp/Kconfig:23:warning: range is invalid

PR #5377 shows that range was for CONFIG_NET_TCP_RCV_SCALE. And description of that config
also says the same range.
This commit moves it to proper config to resolve the warning.

Signed-off-by: sunghan-chang <sh924.chang@samsung.com>