os/binfmt: Refactor code for usage of uheap and mpu_regs from parent tcb
We use the uheap pointer to verify if the parent task is a kernel side task
and accordingly allocate stack for the child. However, since we have now
blocked the creation of task and pthread from kernel side, this check is not
required. We also modify the loader code to copy the mpu_regs into the newtcb
after the child task has been created. So, we have removed the need to set
the uheap and mpu_regs in the binary manager tcb (which is parent for apps).

Signed-off-by: Kishore S N <kishore.sn@samsung.com>