mm_check_heap_corruption.c : update the conditionals for the use of abort mode
This PR is for addressing the comments in PR #5705.
The abort mode flag is used in case of board crash. This flag is used in the mfdbg
definition in os/include/debug.h for mfdbg conditionals. So changing the conditional
for assignment of the board status for printing the heap information.