log_dump.c: Fix hang issue in log dump module
If log dump is enabled, and we run heapinfo command, then
there is a hang issue. This is because, the heapinfo command
locks the kernel heap. But, at the same time heapinfo logs
are sent to log dump module which tries to allocate buffer.
Since, the heap is already locked by the heapinfo command,
it results in a lock-up.

To overcome this issue, upon entering logdump api we will check
if the kernel heap is locked before performing the log dump
operations which require memory to be allocated or freed. The
log dump operation will not be performed as long as the heap is
locked. So, any logs printed during this time will not get included
in the log dump.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>