log_dump.c: Fix hang is log dump due to priority mismatch
The log dump compression thread runs at a predefined priority. If a
thread with higher priority tries to perform the log dump operation,
then there can be a lockup since the low priority compression thread
cannot execute, and the high priority app thread will be blocked for
the compression to complete.

To fix this issue, we check if the current thread is having higher
priority than the log dump compression thread, and in such case, the
current logs will be ignored and not included into the log dump. This
will avoid lock up at the cost of missing some logs from the dump.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>