arch/arm/common: correct the range to check stack_color
The range of setting stack_color or checking stack_color is not the same.

In set stack_color:
	start: (stackbase + 3) & ~3  //align up
	end  : (stackbase + stacksize) & ~3  //align down
In check stack_color:
	start: (stackbase) & ~3  //align down
	end  : (stackbase + stacksize + 3) & ~3  //align up

So, It can check the range that is not set to stack_color. and it can make incorrect the stack usage log.

Therefore, This commit correct range of check stack color.
And it modifies to use STACK_ALIGN_UP, STACK_ALIGN_DOWN in setting stack_color and checking stack_color.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>