mm_heap/mm_initialize.c: set total_alloc_size to include first and last node
total_alloc_size is being initialized to zero and is being updated only during
mm related calls (malloc, free etc). So, it will not include the size of the
guard nodes. If we check mallinfo implementation, the allocated size includes
the size of guard nodes. So, initialize it to proper value.

before fix :
[tc_umm_heap_get_heap_free_size] FAIL [Line : 485] umm_get_heap_free_size : Values (free_size == 0x736c0) and (info.fordblks == 0x736a0) are not equal

after fix :
[tc_umm_heap_get_heap_free_size] PASS
.
.
.########## Kernel TC End [PASS : 415, FAIL : 0] ##########

Signed-off-by: Abhishek Akkabathula <a.akkabathul@samsung.com>