os: Use enter/leave_critical section for SMP
TizenRT originally makes use of irqsave and irqrestore API for
to provide critical sections. However, when SMP is implemented, we
need to consider the irq and lock status on multiple cores. This is
implemented in enter/leave_critical_section APIs. When the SMP
configuration is enabled, we need to use these new API's instead of
irqsave/restore.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>