os: Fix to include Samsung header in hex, srec binary
The $BIN.bin file is used to convert to hex or srec binary.
But if binary name is changed in MAKE_BOARD_SPECIFIC_BIN, This use binary that
is not board specific binary, also the using binary is without samsung header.

So This commit changes to use binary with samsung header when convert bin to hex or srec.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>