debug: Add assertdbg feature
Add assert debug feature.
The assertdbg is used for enhanced security for assert debug feature.
It outputs syslog same as lldbg, but it checks security level before output.

This assertdbg feature is enabled from CONFIG_SECURE_DEBUG_ASSERT.
if CONFIG_SECURE_DEBUG_ASSERT is disabled, it's operation is exactly the same as lldbg.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>