log_dump: change to not compress operation during sched_lock
when the log is outputted while the scheduled lock is called, hang can be occur.
For this, applied a8822cb commit. but This is breaking the intention of blocking context
switching using sched_lock in the application, cpu is forcibly handed over to another thread.

So, This commit adds so that there is no compress operation when the sched_lock is stuck.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>