platform/gnu, libxx: select config appropriately for static constructors init
In the case of flat, CONFIG_HAVE_CXXINITIALIZE is required and in the
case of loadable, CONFIG_BINFMT_CONSTRUCTORS is required for cpp static
constructors initialization. Hence, select them appropriately when cpp is
enabled. Users need not select them specially as from now on they are enabled
by default.

Make CONFIG_HAVE_CXXINITIALIZE dependent on BUILD_FLAT so users will not
enable it in loadable configs by mistake

Signed-off-by: Abhishek Akkabathula <a.akkabathul@samsung.com>