kernel/log_dump: refactor log_dump
log_dump still have a lot of issues.
So, this commit rewrite the log_dump module in order to increase the readability and memory efficiency.
The previous module had to wait until it was compressed when the buffer was full.
In the reconstructed module, threads that write logs just write and leave until one block of buffer is full.
When a block of buffer is full, just request compress message to log_dump_manager.
Log writing thread should not wait until entire buffer is full.

Signed-off-by: zambbo <rhkrqnwk1@naver.com>