mm/mm_heap, mem_leak_checker: fix heap corruption hang.
The chunk of heap can know the address of the next chunk through `node = node + node->size`.
Using this, all chunks of the heap can be traversed in the forward direction.

If the heap is corrupted and the node->size becomes 0, It can't move to the next chunk, and the iteration turns infinite.
Also, It is Locked the heap during the heap traversing. As a result, it becomes deadlock.

Therefore, Add to check the node size and it to cause ASSERT if it is zero during traversing forward direction.

hang reproduce code:
```
	int *val = (int*)malloc(sizeof(int) * 100);
	val = val - 1;
	*val = 0;

	mallinfo();
```
Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>