arch/arm/amebad/Kconfig: change spi default value to n, rename config
The config related to SPI is setting to default y, but spi is not required as a mandatory.
So, change the default value of AMEBAD SPI to n.

And This commit changes the name of config to CONFIG_AMEBAD_XXX to know that it is a config regarding specific AMEBAD board.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>