arch/arm/armv8-m/up_blocktask: Correct invaild tcb->tz_context check
If thread is in the trustzone, (TCB)->tz_context is allocated.
And when context switching while this thread is operating, saves the context related
to the trustzone in (TCB)->tz_context.

Likewise in up_block_task, the tz_context of rtcb(running thread) must be checked
and the context of the trustzone must be stored, but the tz_context of the wrong TCB
is checked.

Therefore, Correct to check the tz_context of running tcb.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>