os/arch, os/binfmt: Implement loadable apps memory protection with MMU
When MMU is enabled, we will use L1 and L2 page table to provide memory
proctection to loadable apps. At the time of loading, the page tables are
created for each app. At the time of context switch, we change the page table
bsae register to point to the new app's page table. In case of common binary,
the page tables will be created as part of kernel page tables and will be
used by each app.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>