include/tinyara/securit_level: add CHECK_SECURE_PERMISSION macro
To check security level is uesd many location in kernel feature.
But when we check security level, we must check CONFIG_SECURITY_LEVEL before.
So, It is not readable as blow.

```
#ifdef CONFIG_SECURITY_LEVEL
#include <tinyara/security_level.h>
#endif

#ifdef CONFIG_SECURITY_LEVEL
	if (get_security_level() == LOW_SECURITY_LEVEL)
#endif
	{
```

Therefore, This commit adds CHECK_SECURE_PERMISSION macro that returns whether security level is low,
or if CONFIG_SECURITY_LEVEL is disabled, returns true.

Also, This commit changes security_level.h to be able to include at any time, and change
security level APIs to be defined only when CONFIG_SECURITY_LEVEL is enabled.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>