os/arch/armv7-a: Fix prefetch abort when CONFIG_OPTIMIZE_APP_RELOAD_TIME=n
When CONFIG_OPTIMIZE_APP_RELOAD_TIME is disabled, then the memory permission
is set to RW-X so that we can execute the code in this region. However, this
case of RW-X was not handled in the mmu setting APIs. Hence the permission was
set as RW-NX which resulted in prefetch abort. Add new RW-X setting in mmu to
fix the issue.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>