os: Remove assertdbg and replace CHECK_SECURE_PERMISSION
we have two methods for controlling the output of assert logs based on security levels: 'assertdbg()'
and 'CHECK_SECURE_PERMISSION()'. However If disable CONFIG_SECURE_DEBUG_ASSERT, It do not affect logs
controlled by 'CHECK_SECURE_PERMISSION()', leading to potential confusion.

To simplify the codebase and align security level checks, we have removed 'assertdbg' and exclusively
rely on 'CHECK_SECURE_PERMISSION()' to manage assert log outputs. But If add it to all assert log outputs,
It is not desirable. So, we collect the output log in units of fault and check it at once.

This change enhances clarity and maintains consistent behavior with security level configurations.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>