os/board/rtl8730e: Fix heap corruption during app reload
Current implementation reserves some space at the end of BDRAM region
for setting up page tables. This reserved space is getting included at
the end of the kernel heap. However, when we reload the apps, we clear
the page tables completely. As a result the last node information of the
heap gets corrupted. To avoid this, we modify the heap so that it does
not include the areas reserved for page tables.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>