mm/mm_heap: fix crash when CONFIG_DEBUG_MM_INFO is eanbled
When CONFIG_DEBUG_MM_INFO, dbg is used to print few logs during
task clean up stage. This causes issue as during task clean up
the stdout stream realted semaphore is uninitialized in the func
lib_stream_release , but the same sem is used later for dbg logs.
Thus, this results in assert failure in lib_filesem.c as the error
number we get when taking sem on an invalid sem is EINVAL

So, use mllvdbg instead of mvdbg

Signed-off-by: Abhishek Akkabathula <a.akkabathul@samsung.com>