framework/include/security : Increase secure storage data slot size
There is a slot that stores AP information, but the required size has increased while storing multi-AP information.
Since the number of slots is limited, it is more efficient to store the data by increasing the size of existing slots instead of dividing it into several slots.
Therefore, this commit increase data slot size from 256 to 4096.
SECURITY_MAX_CERT_BUF is also changed to 4096 for the identity with trt_se.h (product code)