framework/include/security : Increase secure storage data slot size
There is a slot that stores AP information, but the required size has increased while storing multi-AP information.
Since the number of slots is limited, it is more efficient to store the data by increasing the size of existing slots instead of dividing it into several slots.
Therefore, this commit increase data slot size from 256 to 4064.
Maximum size of a slot is 4KB,
but when data is stored in one slot, 32 bytes of tag is added.
Therefore, user can store 4064 bytes of data.