os/kernel: Fix assert fail during app loading with SMP
When SMP is enabled, and app loading thread is exiting, then there
is an assert fail since the lockcount and the sched lock state of the
tcb are not proper. Since we are incrementing the lockcount before
task terminate, we need to make changes accordingly to the lock set
before and after task termination.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>