arch/armv7-a: Fix prefetch abort in loadable apps with SMP
There is a prefetch abort when app1 is loaded and tries to run
first instruction. This is happening because the page table for
the app is not setup for second core before activating the app's
task.

The app1 startup thread is running on core0, but the app1 task
is setup to run on core1. In this case, the context switch api
is not called for core0, but the cpu pause/resume apis on core1
perform the context switching. So, we need to setup the TCBR
for core1 in the cpu pause/resume sequence before thea app1
task starts to run.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>