os/kernel/task: Fix crash in sched_unlock for flat apps SMP
When running smp test app using flat apps config, there is an
assert fail due to improper values of lockcount and g_cpu_lockset.
This is because we are incrementing lockcount in prepare_exit API
without the corresponding setting of g_cpu_lockset. To fix this issue
we modify prepare_exit API to increment and decrement lockcount along
with set and clear of g_cpu_lockset.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>