os/kernel: Fix assert fail in sched_unlock during kernel tc
When running kernel tc there is an assert fail in sched_unlock.
In this case, sched_unlock is called when the lockcount is already
zero. As a result, checking for the g_cpuschedlock and g_cpulockset
is causing an assert.

But, since the lockcount is already zero, this check is wrong. The
check needs to be done only when sched_unlock is called with a
non-zero lockcount value.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>