os/kernel/task: Fix assert failure in sched unlock during wm_test app
When we run wm_test scan command, there is an assert failure in
sched_unlock. This is because the current tcb has a non-zero lock count,
whereas the g_cpu_lockset variable is zero.

The fix contains 2 parts:
1. In prepare_exit API, perform set and clear of the lockset only if it
is not already set. If lockset is already set by another process, then the
exiting process must not clear it.

2. In task_exit API, when we try to get the new task at the head of the
readytorun list, we must use this_cpu() to get the current cpu. This is
because during task_exit, the task might get blocked and switched to a
new cpu. So, while trying to get the new task, we need to make sure that
we are using the correct cpu.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>