apps/examples/testcase: Fix mqueue kernel tc failure with SMP
When SMP config is enabled, the mqueue kernel tc is failing because
it makes use of multiple pthreads of different priority to send and
recieve messages. However, when smp is enabled, having different
priority does not guarantee the order of execution becuase the pthreads
are getting scheduled on multiple cpus. To fix this issue, set the affinity
of all pthreads in the mqueue test to cpu 0.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>