os/kernel/task, os/arch/arm/src/common: Fix stack overflow issue when printing TCB list in assert logs
when printing the assert information, the size of stack used for each
TCB is printed wrong and showing stack overflow. This is because when
we are checking the TCB stack, we check from stack_alloc_ptr instead of
stack_base_ptr. So, stack_alloc_ptr points to TLS Data, Arguments and
then available stack whereas stack_base_ptr directly points to available
stack.

Signed-off-by: Neelkumar Patel <neel.patel@samsung.com>