kernel/preference: Return if registered callbacks are not exisiting or sending callback messages are done.
If a thread registers callback, it is invoked when value of the given key in the preference changes.
This callback operates based on signals and message queues.
If there is a registered callback, it sends data to the created message queue and sends a signal then a handler that receives the callback deletes this created mq.
So just return if registered callbacks are not exisiting or sending callback messages are done.