os: Add tick suppression operation in PM
Periodic tick interrupts can be suppressed to keep the CPU idle for a longer period of time.
Tick suppression requires that missed ticks while the core is in an idle state be compensated
for when the core becomes active.

Also, it should be possible to specify which one should compensate for the missing tick within the tizenrt OS, not the bsp code.

Therefore, Add a code that compensates for missing ticks due to CPU idle state in the pm code.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>