kernel/log_dump: Add for detecting device hang occurrence
Currently, the log dump structure prevents logging output until compression is complete, which can cause device hangs.

Although exceptions have already been processed in several cases, there may be other exceptions.
Also, when a hang occurs, the log is not printed, making it difficult to determine the cause.

Previous device hangs generated in the log dump occurred because a simple busy wait was performed
with a while statement to wait for compression.
In other words, if you cannot get out of the while loop, a device hang may have occurred.

Therefore, if the while statement is not exited within a certain time,
it calls PANIC to notify that a device hang has occurred and debugs by printing the assert log.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>