apps/shell/tash_command: Add pm suspend while executing SYNC cmd
We check the information or status of the program by using the TASH SYNC command.
TASH is only used for debugging and testing, so it does not affect power consumption in actual operation.
Therefore, for proper debugging and testing, we apply pm suspend to the TASH SYNC command to ensure operation.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>