configs/rtl8730e: Add temp defconfig for loadable apps with SMP
This is a temp defconfig in which SMP is enabled with loadable apps.
It will be used for development and testing of PM with SMP. Once, it
is finalised, then SMP will be enabled by default in flat and loadable
apps configs.

This commit has also temporarily commented out the call to PM api
arm_gic_raise_softirq() to fix a build error. This also needs to be
reverted once PM implementation is completed and verified.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>