os/include/tinyara/mm: Fix crash in loadable smp when there is single app
When a single app is present then the app id will always be 1.
When we are allocating memory from user side, then we are sure that
the allocation has to be done from this single app heap only. All
kernel side allocations will be done by the kernel MM. So, we hardcode
the value of BASE_HEAP to use app1 for all user allocations.

Signed-off-by: Neelkumar Patel <neel.patel@samsung.com>