os/tools/check_package_header.py: Fix non-signing build error
SIGNING_SIZE is the header size of signing user binary.
But if CONFIG_USER_SIGN_PREPEND_SIZE is not set, SIGNING_SIZE value is 'None'.
so it causes build error when directly convert SIGNING_SIZE type from string to int.
This commit adds the process when the SIGNING_SIZE is none.