os/kernel/wdog: Add wdog wakeup source feature
In the tick suppression and the CPU is in idle, the CPU can be waken up
by a specific watchdog. When this config is enabled, we can specify the wdog
to use as a wakeup source. To wake up the CPU with the specified wdog,
need to register the expiration time of the wdog with the hw timer and wake up
the CPU with the hw timer interrupt.

This feature can be used by enabling CONFIG_SCHED_WAKEUPSOURCE.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>