pm: Remove pm_timer and replace to wd_setwakeupsource
The pm_timer has almost the same code structure as the wdog,
but the application cannot directly use the wdog.
Therefore, it is necessary to add a code for the user to use pm_timer,
but this will have the same structure as existing codes and thus it would be difficult to maintain.

Thus, the function of designating wake-up (wd_setwakeupsource) in the wdog is replaced, and the pm_timer is removed.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>