os: fix pm wakeup hang issue
The pm_wakehandler is currently experiencing a hang.
This is caused by the IDLE thread giving up the CPU while executing the pm_wakehandler.
The pm_wakehandler is executed by the IDLE thread after the main core has awakened and the tick timer is still stopped.
Therefore, if the IDLE thread gives up the CPU while executing the pm_wakehandler, it cannot regain the CPU without another
thread voluntarily relinquishing it, which prevents the tick timer from restarting.
This is because the tick timer is stopped and the RR scheduling does not work.

The pm_wakehandler includes the wd_expire operation. However, since the wd expire handler can include context switching
operations such as sem_post or signals, the IDLE thread may give up the CPU.

Therefore, Change to so that, the suppressed tick is corrected in the pm_wakehandler and the wd_exipre is modified to be executed in the next tick ISR.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>