kmm_realloc: fix crash issue case of oldmem is NULL
if oldmem argument is NULL, we should call the malloc(). It is posix rule.
But now, if oldmem is NULL, mm_get_heap returns NULL. So we get crash.

So, Add checking NULL.
And Change to so that, if oldmem is out of heap address, we can get a assertion.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>