app/examples/testcase/le_tc/kernel : Fix kernel_tc testcase for rtl8721csm board
Make required changes to resolve the failed kernel testcase.
There was two tc is failing: First in tc_task.c and 2nd in tc_timer.c
In tc_task.c: it was working only in case of when SMP was enabled,now it is working in both enable and disable cases.
And in tc_timer.c: it is a minor change to check the set time(ie: st_timer_spec_set.it_value.tv_sec) be greater than the get time(st_timer_spec_get.it_value.tv_sec).

Signed-off-by: gaurav-15 <gaurav.15@partner.samsung.com>