os: Add support for Asymmetric Multi Processing (AMP) in kernel
Currently during boot up, we set the affinity of the idle thread to run on any core. As a result,
all the threads that are created later can run either on any core. For AMP, we modify
our idle thread to run only on core 0. As a result, all threads (child threads of idle thread)
will be forced to run only on core 0. The user may later change the affinity of any task and make
it run on other cores.

Signed-off-by: neel-samsung <neel.patel@samsung.com>