apps/shell/tash_command: fix tash_help character breakage issue.
If a string longer than 15 characters is registered as a tash command like 'testtest_testsetup',
there is an issue where the end of the string gets corrupted when the 'help' command is executed,
resulting in output like 'testtest_testse.Y9'.

Since tash only copies the first 15 characters when registering a command,
the last character in the buffer is being printed out as garbage.

To fix this, Add a null character '\0' at the 16th position so that even if a command with
more than 15 characters is registered, the output value won't get corrupted.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>