os/kernel/pthread: Fix issue during pthread_create with affinity
If pthread_create is called with non-zero value for affinity, then
we need to check if SMP is enabled. We need to take the affinity value
only if SMP enabled. However, if SMP is disabled, the pthread_create
API must not fail, so that the app layer can still use the pthread API
irrespective of SMP config in kernel. So, instead of returning an error
code at this point, we just print an error message and continue
with the pthread_create API.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>