ai-soc/ndp120: fix wrong argument value in extract audio
when "SYNTIANT_NDP_ERROR_DATA_REREAD" error occurs, zero is
returned in the argument extract size as no bytes are extracted.
However, in the next iteration, zero is passed as extract size,
causing to extract the entire audio tank, hence corrupting the
heap. This exactly matches the crash we observed as the heap
corrupted (around 60KB) matches the size of the audio tank (60KB)

Signed-off-by: Abhishek Akkabathula <a.akkabathul@samsung.com>