pm_domain_register: fix heap corruption issue.
When creating a new domain, it allocs as much memory as the length of the domain string.
However, the domain is copied to the allocated memory as long as the domain length + 1
Because to store a space character.

It is breaking the next chunk header of the allocated memory, resuting in a corrupted heap.

Therefor, Fix to allocate length + 1.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>