lib/libc/syslog: Fix log mix issue between printf and dbg
Printf is using lib_stdoutstream_s and dbg is using lib_rawoutstream_s.
Due to this, the logs from printf are buffered in stdio buffer. But dbg
logs are sent to uart one character at a time.

Since the stdout stream is duplicated for each task, taking a semaphore on
stdout is not providing exclusive access. This is resulting in mix up of
logs.

To fix this, we modify the dbg to use buffered lib_stdoutstream_s.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>