os/drivers/serial: avoid board sleep in UART RX
This commit make use of PM APIs to avoid board sleep when RX buffer is empty
and application thread waiting for it to get filled. Now, when we get
UART IRQ, we suspend the board sleep and resume it after the
application thread consumed all the RX buffer data. This will ensure
board will remain active whenever there is data in RX buffer.