os/arch/arm/src/armv7-a: Add logic for cpu pause
If a cpu is already paused and again up_cpu_pause is called,
then it might result in a lock up due to waiting on semaphores.
So, in up_cpu_pause we check whether a pause reques is pending
or whether the cpu is already paused. Then we return without
sending another pause request.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>