os: Add pm_start api
This commit enable applications to start power
management system. This API should be called
when board is ready to transition into low
power state. This API helps PM to remove the
use of PM_IDLE_DOMAIN outside the scope of PM.

pm_start - This API meant to be called during app
initialization to enable low power state transition.
In PM side we can reduce some of the costlier operation
based on the is_running flag. In the event of crash
and app reloading, this API can be used for reinitialization.

pm_resume - This API removes pm state transition lock for
specific domain. It has domain information, which helps PM
to do domain specific operation, like "when LCD domain get
resume, do backlight dimming". This should be used in pair
with pm_suspend API.

pm_suspend and pm_resume are like locks, which should be used
in pairs to prevent board sleep in critical section. The current
usage pm_suspend(0) in OS and pm_resume(0) in application, to prevent
board sleep before app initialization is not good design. So, we kept
PM (OFF) by default and pm_start API will do PM (ON) when application
is ready to do power saving.