drivers/lcd: Add to prevent entering board sleep mode when backlight is on
When the LCD backlight is on, it indicates that the user is currently using the device, so it should not enter sleep mode.
And Since the backlight consumes the most power, it is not meaningful to put the board in sleep mode.

Therefore, Add to prevent entering board sleep if backlight is on

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>