input: Change touchscreen event structure
A specific touchscreen IC reads data immediately when an interrupt occurs,
and it should generate a Touch Event only if that interrupt is touch data.

The current structure gives a notification to the user app upon IRQ occurrence
and waits for the user thread to read the touch data.
Moreover, even in cases where it's not touch data, the user thread still reads the data.

Therefore, Modify the structure so that the data can be directly read from the touchscreen IC
 as soon as the IRQ occurs, generating notifications to users only when it's a touch event.

The change involves reading from /dev/touch0 in the application as a "struct touch_point_s"
instead of a "struct touch_sample_s".

Reading as a "struct touch_sample_s" restricts the number of touch points that can be read
at once to the value of TOUCH_MAX_POINTS set in the driver.

We should not limit the size of touch point data that can be read at once in the application
based on the driver's restrictions.
Therefore, we modify it so that the application can read the desired amount of data corresponding
to the array size of the "struct touch_point_s".

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>