os/arch/arm/amebasmart: Fix I2C hang issue.
it's calling sem_wait for i2c usage.
The i2c hang issue is occuring if it is using i2c in work queue.

The work queue operation is using signal a lot, and The signal can wakeup a sem blocking with return EINTR errno.
So, we should check if errno is EINTR, then we call sem_wait again.

The i2c hang issue is wrong check condition for calling sem_wait again above case .
The while condition should be ret != OK.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>