drivers/input: Add ist415 touch ic suspend feature.
The touch device suspend feature is changing power mode to low power mode.

In the low power mode, power comsumption is reduced and touch irq is not comming.
But ist415 ic is checking double tap during low power mode, and it notify to apps the double tab event.

In this commit, Add ioctl to control touch ic suspend/resume and example app

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>