os/arch/armv7-a: Add enter_critical_section in up_schedule_sigaction
The up_schedule_sigaction increases the target task's stack frame to execute the sigaction operation,
 allowing the sigdeliver operation to be performed before the running operation.

Since it changes the stack, SP, and PC, it should be protected from IRQ operations including context switching.

If unprotected, a context switch may occur during the change of stack and PC, causing the target task to operate
with incorrect stack SP and resulting in a crash.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>