arch/arm/armv7-a: Modify to printing LR when NULL executing fault
If null is executed, prefetch abort is occuring.
But PC and R15 value is 0x00000000.

For debugging this fault, we need to save LR(R14) and print it in fault log.

Therefore, Add print LR in pretech abort logs.
And if PC is not text address, replace system exception location to LR(R14).

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>