arch/arm: Correct exception location "LR - 4" when PC invalid fault case
If PC becomes invalid causing a fault, we replace the system_exception_location with the address at LR.

However, since the value of LR is the address to execute after returning,
the actual address that caused the fault is the address -4.

Therefore, when replacing LR with exception location, it is corrected to address - 4.
This will help avoid confusion for developers working on debugging applications such as TRAP.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>