os/board/rtl8730e/src/compoent: add critical section to GPIO write bit
    1. With SMP/AMP, there might be race condition when both Core 0 and Core 1 attempt to modify the GPIO simultaneously, causing the the value pin to be abnormal
    2. enter_critical_section is required to ensure that only one core can modify the GPIO register at any given time, we can prevent simultaneous access to the GPIO data register, thus eliminating the race condition.

Test results:
without this PR
LCD product
- Test sequence : booting -> format all -> …
- Original product code + without bixby : Assert can be observed without bixby operation
- Original Product code + with bixby
Can be observed “part_bread: ERROR: Read beyond the end of the partition” before the assert.
So above sentence is used for the verification result

With this PR, Product code + with bixby
No “part_bread: ERROR: Read beyond the end of the partition” can be found.