arch/arm/arm7-a: Add to check condition hotplug in cpu_pause
The device hang occurs when cpu pause is attempted for a CPU in hotplug state.

Therefore, Check the status of the CPU that intends to pause and add an assertion if it is in a hotplug state.
Additionally, since there is a CPU pause in up_assert, it is necessary to return in the abort_mode state
to prevent the assertion from being called redundantly.

Signed-off-by: eunwoo.nam <eunwoo.nam@samsung.com>