os/mm, os/drivers, os/include: Fix mem alloc fail due to dead task stack
When a task or thread is terminated, we place the stack details in a queue
and it will be freed at a later point using the LPWORK thread. However, if
LPWORK is not able to run due to some reason, then we run into "out of mem"
error.

This commit adds code to check and release all the dead task stack space
when we run into "out of mem" scenario.

Signed-off-by: Kishore S N <kishore.sn@samsung.com>